pyproject.toml
[project]
name = "auto-research-on-databricks-serverless"
version = "0.1.0"
description = "Autonomous ML research on Databricks Serverless GPUs: LoRA fine-tuning Qwen 3.5 4B on financial text"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.45.0",
    "peft>=0.13.0",
    "trl>=0.12.0",
    "datasets>=3.0.0",
    "accelerate>=0.34.0",
    "bitsandbytes>=0.44.0",
    "mlflow>=2.16.0",
    "databricks-sdk>=0.36.0",
    "openai>=1.50.0",
]

[project.optional-dependencies]
eval = [
    "anthropic>=0.39.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
